Refactor file structure & delete unused code
