+++ src/components/LinkLogo/LinkLogo.tsx
@@ -1,15 +1,7 @@
 import { IconDefinition, SizeProp } from '@fortawesome/fontawesome-svg-core';
 import { useState } from 'react';
-import { LinkLogoLabel, LinkLogoWrapper } from '../shared/styledComponents-OLD';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
-interface LinkLogoProps {
-  iconName: IconDefinition;
-  link: string;
-  label?: string;
-  newTab?: boolean;
-  size: SizeProp;
-}
+import { LinkLogoLabel, LinkLogoWrapper } from './LinkLogo.styles';
 
 function LinkLogo({ iconName, link, label, newTab, size }: LinkLogoProps) {
   const [isHover, setIsHover] = useState<boolean>(false);
@@ -23,5 +15,13 @@ function LinkLogo({ iconName, link, label, newTab, size }: LinkLogoProps) {
     </a>
   );
 }
+
+type LinkLogoProps = {
+  iconName: IconDefinition;
+  link: string;
+  label?: string;
+  newTab?: boolean;
+  size: SizeProp;
+};
 
 export default LinkLogo;
