+++ src/components/ProjectCard/ProjectCard.tsx
@@ -1,0 +1,3 @@
+function ProjectCard() {}
+
+export default ProjectCard;
