--- src/sections/AboutSection.tsx
@@ -1,48 +1,0 @@
-import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
-import LinkLogo from '../components/LinkLogo';
-import '../styles/globals.css';
-import {
-  HorizontalFlexBox,
-  ParagraphText,
-  Section,
-  SectionTitle,
-} from '../shared/styledComponents-OLD';
-import {
-  ABOUT_ME_PARAGRAPH_1,
-  ABOUT_ME_PARAGRAPH_2,
-  GithubLabel,
-  GithubUrl,
-  LinkedinLabel,
-  LinkedinUrl,
-} from '../constants';
-import { isMobile } from 'react-device-detect';
-
-function AboutSection() {
-  return (
-    <Section className="hidden altBackground">
-      <SectionTitle className="hidden lightThemeText">About Me</SectionTitle>
-      <ParagraphText className="hidden lightThemeText">{ABOUT_ME_PARAGRAPH_1}</ParagraphText>
-      <br />
-      <ParagraphText className="hidden lightThemeText">{ABOUT_ME_PARAGRAPH_2}</ParagraphText>
-      <br />
-      <HorizontalFlexBox className={`hidden ${!isMobile && 'paddedTop smallMarginBottom'}`}>
-        <LinkLogo
-          iconName={faLinkedin}
-          link={LinkedinUrl}
-          label={LinkedinLabel}
-          newTab
-          size={isMobile ? '5x' : '10x'}
-        />
-        <LinkLogo
-          iconName={faGithub}
-          link={GithubUrl}
-          label={GithubLabel}
-          newTab
-          size={isMobile ? '5x' : '10x'}
-        />
-      </HorizontalFlexBox>
-    </Section>
-  );
-}
-
-export default AboutSection;
