+++ src/sections/Contact/ContactSection.tsx
@@ -1,20 +1,14 @@
 import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons';
-import LinkLogo from '../../components/LinkLogo';
-import {
-  ContactSectionText,
-  ContactSectionTitle,
-  ContactSectionWrapper,
-} from './ContactSection.styles';
-import { GithubLabel, GithubUrl, LinkedinLabel, LinkedinUrl } from '../../constants';
+import LinkLogo from '../../components/LinkLogo/LinkLogo';
+import { ContactSectionText, ContactSectionWrapper } from './ContactSection.styles';
+import { GithubUrl, LinkedinUrl } from '../../shared/utils/Constants';
 import { isMobile } from 'react-device-detect';
-import { Flex } from '../../shared/styledComponents';
+import { Flex, SectionTitle } from '../../shared/styles/styledComponents';
 
 function ContactSection() {
-  // TODO: Unimplemented
-
   return (
     <ContactSectionWrapper className="hidden altBackground">
-      <ContactSectionTitle className="lightThemeText">Contact</ContactSectionTitle>
+      <SectionTitle className="lightThemeText">Contact</SectionTitle>
       <ContactSectionText className="lightThemeText">
         I'm always looking to meet people and love to engage with the Tech community, so feel free
         to reach out at the links below!
@@ -23,14 +17,14 @@ function ContactSection() {
         <LinkLogo
           iconName={faLinkedin}
           link={LinkedinUrl}
-          label={LinkedinLabel}
+          label={'LinkedIn'}
           newTab
           size={isMobile ? '5x' : '10x'}
         />
         <LinkLogo
           iconName={faGithub}
           link={GithubUrl}
-          label={GithubLabel}
+          label={'GitHub'}
           newTab
           size={isMobile ? '5x' : '10x'}
         />
