--- src/sections/PortfolioSection.tsx
@@ -1,13 +1,0 @@
-import { Section, SectionTitle } from '../shared/styledComponents-OLD';
-
-function PortfolioSection() {
-  // TODO: Unimplemented
-
-  return (
-    <Section className="hidden altBackground">
-      <SectionTitle className="lightThemeText">Portfolio</SectionTitle>
-    </Section>
-  );
-}
-
-export default PortfolioSection;
